Specific class file for BackupAction
